Factor ConstantRecycler template in its own file.
